split notes tree rows and header actions out of channel screen
